Break out backend tests over multiple files
